requirements.txt
requests
Pillow
matplotlib